refactor: arrange witness io
