+++ yarn-project/acir-simulator/src/execution.ts
@@ -1,25 +1,7 @@
-import { ACVMField, acvm, ACVMWitness, toACVMField, fromACVMField } from './acvm.js';
-import {
-  ARGS_LENGTH,
-  AztecAddress,
-  CallContext,
-  ContractDeploymentData,
-  EMITTED_EVENTS_LENGTH,
-  EthAddress,
-  Fr,
-  L1_MSG_STACK_LENGTH,
-  NEW_COMMITMENTS_LENGTH,
-  NEW_NULLIFIERS_LENGTH,
-  OldTreeRoots,
-  PrivateCircuitPublicInputs,
-  PRIVATE_CALL_STACK_LENGTH,
-  PUBLIC_CALL_STACK_LENGTH,
-  RETURN_VALUES_LENGTH,
-  TxRequest,
-} from '@aztec/circuits.js';
+import { ACVMField, acvm, toACVMField, fromACVMField } from './acvm.js';
+import { AztecAddress, CallContext, EthAddress, Fr, OldTreeRoots, TxRequest } from '@aztec/circuits.js';
 import { DBOracle, PrivateCallStackItem } from './db_oracle.js';
-import { frToAztecAddress, frToBoolean, frToEthAddress, WitnessReader, WitnessWriter } from './witness_io.js';
-import { randomBytes } from '@aztec/foundation';
+import { arrangeInitialWitness, extractPublicInputs, frToAztecAddress } from './witness_io.js';
 import { FunctionAbi } from '@aztec/noir-contracts';
 
 export interface ExecutionPreimages {
@@ -63,7 +45,6 @@ export class Execution {
     return this.runExternalFunction(
       this.entryPointABI,
       this.contractAddress,
-      1, // TODO this comes from contract ABI
       this.request.functionData.functionSelector,
       this.request.args,
       callContext,
@@ -74,13 +55,12 @@ export class Execution {
   private async runExternalFunction(
     abi: FunctionAbi,
     contractAddress: AztecAddress,
-    witnessStartIndex: number,
     functionSelector: number,
     args: Fr[],
     callContext: CallContext,
   ): Promise<ExecutionResult> {
     const acir = Buffer.from(abi.bytecode, 'hex');
-    const initialWitness = this.arrangeInitialWitness(args, callContext, witnessStartIndex);
+    const initialWitness = arrangeInitialWitness(args, callContext, this.request.txContext, this.oldRoots);
     const newNotePreimages: Array<{ preimage: Fr[]; storageSlot: Fr }> = [];
     const newNullifiers: Fr[] = [];
     const nestedExecutionContexts: ExecutionResult[] = [];
@@ -108,7 +88,7 @@ export class Execution {
       },
     });
 
-    const publicInputs = this.extractPublicInputs(partialWitness, acir);
+    const publicInputs = extractPublicInputs(partialWitness, acir);
 
     const callStackItem = new PrivateCallStackItem(contractAddress, functionSelector, publicInputs);
 
@@ -141,82 +121,5 @@ export class Execution {
   private async getSecretKey(contractAddress: AztecAddress, address: ACVMField) {
     const key = await this.db.getSecretKey(contractAddress, frToAztecAddress(fromACVMField(address)));
     return [toACVMField(key)];
-  }
-
-  private arrangeInitialWitness(args: Fr[], callContext: CallContext, witnessStartIndex: number) {
-    const witness: ACVMWitness = new Map();
-
-    const writer = new WitnessWriter(witnessStartIndex, witness);
-
-    writer.writeFieldArray(
-      new Array(ARGS_LENGTH).fill(Fr.fromBuffer(Buffer.alloc(Fr.SIZE_IN_BYTES))).map((value, i) => args[i] || value),
-    );
-
-    writer.writeField(callContext.isContractDeployment);
-    writer.writeField(callContext.isDelegateCall);
-    writer.writeField(callContext.isStaticCall);
-    writer.writeField(callContext.msgSender);
-    writer.writeField(callContext.portalContractAddress);
-    writer.writeField(callContext.storageContractAddress);
-
-    writer.writeField(this.request.txContext.contractDeploymentData.constructorVkHash);
-    writer.writeField(this.request.txContext.contractDeploymentData.contractAddressSalt);
-    writer.writeField(this.request.txContext.contractDeploymentData.functionTreeRoot);
-    writer.writeField(false);
-    writer.writeField(this.request.txContext.contractDeploymentData.portalContractAddress);
-
-    writer.writeField(this.oldRoots.contractTreeRoot);
-    writer.writeField(this.oldRoots.nullifierTreeRoot);
-
-    return witness;
-  }
-
-  private extractPublicInputs(partialWitness: ACVMWitness, acir: Buffer): PrivateCircuitPublicInputs {
-    const witnessReader = new WitnessReader(partialWitness, acir);
-
-    const callContext = new CallContext(
-      frToAztecAddress(witnessReader.readField()),
-      frToAztecAddress(witnessReader.readField()),
-      frToEthAddress(witnessReader.readField()),
-      frToBoolean(witnessReader.readField()),
-      frToBoolean(witnessReader.readField()),
-      frToBoolean(witnessReader.readField()),
-    );
-
-    const args = witnessReader.readFieldArray(ARGS_LENGTH);
-    const returnValues = witnessReader.readFieldArray(RETURN_VALUES_LENGTH);
-    const emittedEvents = witnessReader.readFieldArray(EMITTED_EVENTS_LENGTH);
-    const newCommitments = witnessReader.readFieldArray(NEW_COMMITMENTS_LENGTH);
-    const newNullifiers = witnessReader.readFieldArray(NEW_NULLIFIERS_LENGTH);
-    const privateCallStack = witnessReader.readFieldArray(PRIVATE_CALL_STACK_LENGTH);
-    const publicCallStack = witnessReader.readFieldArray(PUBLIC_CALL_STACK_LENGTH);
-    const l1MsgStack = witnessReader.readFieldArray(L1_MSG_STACK_LENGTH);
-
-    const privateDataTreeRoot = witnessReader.readField();
-    const nullifierTreeRoot = witnessReader.readField();
-    const contractTreeRoot = witnessReader.readField();
-
-    const contractDeploymentData = new ContractDeploymentData(
-      witnessReader.readField(),
-      witnessReader.readField(),
-      witnessReader.readField(),
-      frToEthAddress(witnessReader.readField()),
-    );
-
-    return new PrivateCircuitPublicInputs(
-      callContext,
-      args,
-      returnValues,
-      emittedEvents,
-      newCommitments,
-      newNullifiers,
-      privateCallStack,
-      publicCallStack,
-      l1MsgStack,
-      privateDataTreeRoot,
-      nullifierTreeRoot,
-      contractTreeRoot,
-      contractDeploymentData,
-    );
   }
 }
